Add Tag model
This model is designed to store the genre tags we're fetching from
Musicbrainz. Not all MB genre tags have a disambiguation so we've made
that field optional. We also envision using this tags model for tags
that are not sourced from Musicbrainz (e.g. user-generated, or
location tags) so we've made musicbrainz_id optional too.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>