Allow artists to tag their album
This commit adds a multi-select box to the album edit page, and then
displays the chosen tags in alphabetical order on the album show page.

The standard HTML multi-select isn't particularly usable with many
tags, but it works well enough as a default. We'll enhance it with
some JS in a subsequent commit.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>