Enhance tag select box with tom-select.js
This adds a more usable interface for multiple tag selection using
tom-select.js[1]. For now, I've set `create` to `false` as we'd have
to do some additional work to allow users to add their own tags. The
`maxOptions: null` option ensures that all the tags are shown in the
dropdown.

[1] https://tom-select.js.org