Avoid N+1 queries in `ArtistsController#index`
I've had to use `includes` in the template rather than in the controller
action to avoid the following bullet error when the atom feed is
rendered:

AVOID eager loading detected
  Artist => [:profile_picture_attachment]
  Remove from your query: .includes([:profile_picture_attachment])

This is because the atom feed doesn't render the profile picture so
bullet correctly detects that the eager loading isn't required.

I can see an argument for including the profile picture in the atom feed
but I'll think about/address that separately.