Avoid N+1 queries in `AlbumsController#edit`
Note that I can't use `includes` in `#set_album` because bullet
complains that the eager loading isn't required by some of the other
actions that use `#set_album`. I still have to call `#set_album` in
`#edit` so that the `authorize_album` before filter works as expected.
This isn't ideal so I'd like to revisit this to see whether there's a
better way of making this work.