Disable Bullet's check for unused eager loading
I've run into a couple of instances where this check appears to be
incompatible with the check for N+1 queries.

For example: on the artists edit page it complains that cover_image
should be included to avoid executing multiple queries to retrieve the
album cover images if the artist has multiple albums. The fix is to do
something like `Artist.includes(albums: :cover_attachment)`. But if you
do that then it complains that albums is being unnecessarily loaded.