Exempt `Track#transcode` from Prosopite's N+1 detection
I was seeing this problem when editing the artist name in development.
The artist had several albums, each of which had several tracks, and
editing the name triggered the retranscoding of the tracks[1].

The problem appears to be that ActiveJob/SolidQueue executes a SQL select
query after each job has been enqueued. Executing `Track#transcode`
enqueues 3 jobs (one for each of the `Transcode#format` values) and so
Prosopite complains that it's detected N+1 queries:

N+1 queries detected: (Prosopite::NPlusOneQueriesError)
  SELECT "solid_queue_jobs".* FROM "solid_queue_jobs" WHERE "solid_queue_jobs"."id" = $1 LIMIT $2
  SELECT "solid_queue_jobs".* FROM "solid_queue_jobs" WHERE "solid_queue_jobs"."id" = $1 LIMIT $2
  SELECT "solid_queue_jobs".* FROM "solid_queue_jobs" WHERE "solid_queue_jobs"."id" = $1 LIMIT $2
Call stack:
  app/models/track.rb:44:in 'block in Track#transcode'
  ...

I can't think of a good way of avoiding this because there's no direct
association between the `Track` and the `TranscodeJob`s that it creates,
so I can't eager load them. For now I'm configuring Prosopite to allow
N+1 queries for this method in development.

[1]: https://github.com/freerange/jam-coop/blob/45694ffb00e0dac848ae9345c6166d1364afb8ce/app/models/artist.rb#L24