Add beginnings of player page for OpenGraph/oEmbed
I've added this action in a separate controller rather than adding an
action to the `AlbumsController`, because it feels like a different
concern and the access requirements are different.

I've nested the new route under the artists & albums resources so that
it's easy to construct from an an album URL and we can use similar
parameter parsing / model lookup as in the `AlbumsController`.

I'm hoping that this approach means it'll be relatively easy to support
using the `PlayerController` for a single track if we decide to do that.