Add album artwork to player page
I've set the image dimensions to match the ones I'm planning to set
using the "twitter:player:height" and "twitter:player:width" OpenGraph
meta tags. These are the values used by SoundCloud, e.g. here [1].

This might turn out not to be the best idea, but it'll do for now.

[1]: https://soundcloud.com/brenvik/sets/conquest-ep