Wire up Stimulus player controller on player page
This adds the least markup that I could to work with the current
Stimulus player controller that still results in a functional player.

The only CSS classes I've added were ones that I felt I needed for the
player to be at least somewhat useable, e.g. the colours and size of the
progress bar / scrubber.

Note that the `audio` element is now hidden by default, because it's
functionality is provided by other elements.

For now I haven't added any tests to check the player JavaScript
functionality. This seems a low risk given we have such tests for the
album page. However, we might want to revisit this decision.