Use Rails.configuration shortcut
These lines [1] delegate `Rails.configuration` to
`Rails.application.config`, so we might as well use the shorter version.

Note that I haven't updated some other initializers because they were
generated by the `rails new` command and changing them would probably
make it harder to see the wool for the trees when upgrading.

I think there's a case to update the templates in rails itself, but
that's a job for another day!

[1]: https://github.com/rails/rails/blob/b0c813bc7b61c71dd21ee3a6c6210f6d14030f71/railties/lib/rails.rb#L51-L54