Avoid duplication in CDN initializer
Previously we were explicitly setting the non-production
default_url_options, but they were the same as those already set by
`Rails.application.default_url_options` in the development environment.
By using the latter value instead of the former, we can remove some
duplication.

Note that this means the `Rails.configuration.cdn_url_options` will now
be set to `{ host: 'example.com' }` in the test environment instead of
`{ host: 'localhost', port: 3000 }` as it was previously. However, I
don't think this is a problem.