Allow overriding of base URL in dev & prod
This allows the base URL of the app to be overridden using a new
`BASE_URL` environment variable. I've only implemented this in
development and production environments, because it doesn't seem useful
in the test environment.

I've added the host derived from the base URL to the `config.hosts`
whitelist in development, because otherwise if the base URL is set to a
non-localhost domain, host authorization will fail.