Allow user to unsubscribe from newsletter using link
If a user clicks on the "unsubscribe" link in our newsletter (that postmark adds by
default to emails sent by the "broadcast" stream), we receive data
like `webhook-data.json` in the webhook payload. In this case we
don't want to set `sending_suppressed_at` on the User record as this
will then prevent them from receiving transactional emails. Instead we
set the `opt_in_to_newsletter` boolean to false.

Note that users can also change their mind in the subscription page
postmark provides, so we also need to cope with `SuppressSending`
being `false`.