Fix CDN URLs in system tests
It turns out that my note in this commit note [1] was incorrect!

After that commit, CDN URLs had their host set to 'example.com' which
meant that ActiveStorage attachments were not rendered in system tests
and we've seen some instances of system test failing because: "there are
still pending connections" to example.com.

However, this has highlighted a problem that has existed since the
`cdn_url` route helper was introduced. Prior to aforementioned commit
above, CDN URLs had their host and port set to 'localhost:3000'. While
this works fine in development, it didn't work in the system tests,
because Capybara runs the app on localhost with a random port. This
meant that ActiveStorage attachments were not rendered in system tests.
I'm guessing that we saw fewer (no?) system test failures because at
least the host was local to the machine.

Anyway, the way to fix this properly is not to set any CDN URL options
for CDN URL routes in the test environment. This means the URLs are all
relative and thus work OK with the host and port used by Capybara.

[1]: https://github.com/freerange/jam-coop/commit/6c033d88c30de05e488f3869754e6bb7536a4850