Refactor StripeWebhookEventsControllerTest
I want to make it easier to test that the controller can handle another
type of event. The new approach builds just enough of a Stripe event to
satisfy the controller and generates a real Stripe signature for the
request.

Note that I've had to set the Stripe `endpoint_secret` to a non-empty
value in order to avoid an error in the test.