Fix default_url_options in system tests
This doesn't matter for most system tests, because we use the `*_path`
route helpers. However, I'm planning to add a system test which needs to
use the `*_url` route helpers and that won't work unless we make this
change.

The `verify_email_url` implementation in `SigningUpTest` &
`CollectionTest` wasn't robust to URLs with port numbers and now that
`default_url_options` are set correctly we don't need to change the
host. The new implementation is a bit more robust and I've reduced the
duplication by extracting it into `ApplicationSystemTestCase`.