Use Newsletter#body to set the email body
We're rendering the body as markdown (without the tailwind classes) in
the HTML version of the email. We just use the raw markdown as the
body of the plain text email.

I've decided to retain a "Feedback?" footer to be added to all
newsletter emails.

I've also added some dummy content so that the mail preview[1] is
useful.

[1] http://localhost:3000/rails/mailers/broadcast_mailer/newsletter.html