Add Newsletter#send_as_email
This method is designed to be run from the rails console in the first
instance and will send a newsletter to the correct recipients.

We want to send the newsletter to all Users with verified email
addresses, who have not opted out of receiving the newsletter and
whose email addresses are not supressed for other reasons (e.g. bounce
etc).

We also want to send the newsletter to the people who previously
registered interest, confirmed their email address and who are not
suppressed.

We don't have to worry about duplicate emails between the two lists at
the moment because I ensured that anyone on the Interest list who is
also a User was removed from the Interest list by running:

    signed_up = []
    Interest.find_each {|i| signed_up << i if User.exists?(email: i.email)}
    signed_up.each(&:destroy)

in production.