Associate purchase with Stripe Connect account
If a purchase is made from a user who has linked a Stripe account and
that account is accepting payments, when we receive the
"checkout.session.completed" webhook event from Stripe, it should
include a reference to the corresponding payment intent. And that intent
should have `transfer_data.destination` set to the Stripe account
identifier.

When we receive one of these webhook events, we now lookup the payment
intent and pass the destination account identifier to
`PurchaseCompleteJob` so that it can try to associate the `Purchase`
with the relevant `StripeConnectAccount. If a destination account
identifier is provided, but we can't find it in the database, an error
is reported to Rollbar and the job continues to execute as usual, i.e.
the emails are still sent out.

The idea is that we can use this to filter out the purchases for which
we've already made payments via Stripe Connect, so we don't accidentally
make duplicate manual payments via Wise.