Reorganize rake tasks
Previously we enhanced the `test` task with the `htmlbeautifier` &
`rubocop` tasks as prerequisites (running *before* the `test` task) and
with the `test:system` task as an action (running *after* the `test`
task). Previously we were also relying on built-in Rails behaviour which
adds the `test` task as a prerequisite to the `default` task [1].

One downside of this approach was that it changed the default behaviour
of the `test` task which would probably have been surprising to Rails
developers, e.g. running `rails test` normally does *not* run the system
tests.

Now we're clearing the prerequisites & actions of the `default` task and
adding `htmlbeautifier`, `rubocop`, `test` & `test:system` tasks as
prerequisites of the `default` rake task. This means we can still invoke
all those tasks by running `rake` to run the `default` rake task, but we
can also e.g. run just the system tests by running `rake test:system`. I
think this is slightly less surprising.

[1]: https://github.com/rails/rails/blob/90a1eaa1b30ba1f2d524e197460e549c03cf5698/railties/lib/rails/test_unit/testing.rake#L6