Run htmlbeautifier in CI build
Previously the `htmlbeautifier` rake task was running locally when
running `rake` or `rake test` or `rails test`, but it was *not* being
run in CI builds, because we were running the `rubocop` task in a
separate `linters` job and the `test:all` task in the `tests` job. The
latter just runs all the tests in the `tests` directory [1]; it doesn't
run both the `test` & `test:system` tasks as you might expect. I mention
that because previously the `htmlbeautifier` task was a prerequisite of
the `test` task.

Now we're explicitly running the `htmlbeautifier` task as part of the
`linters` job in the CI build just before we run the `rubocop` rake
task.

[1]: https://github.com/rails/rails/blob/90a1eaa1b30ba1f2d524e197460e549c03cf5698/railties/lib/rails/commands/test/test_command.rb#L44-L47