Configure Dependabot's cooldown option
So that Dependabot waits 7 days before opening PRs for version updates
to dependencies[1]. In an attempt to reduce the risk of supply chain
attacks[2].

Note that it's also possible to configure the cooldown differently for
major, minor and patch releases[3] but I think that's probably overkill
at the moment.

Thanks to @floehopper for highlighting this option. I hadn't seen it
before.

[1]: https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#cooldown-
[2]: https://blog.yossarian.net/2025/11/21/We-should-all-be-using-dependency-cooldowns
[3]: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/optimizing-pr-creation-version-updates#setting-up-a-cooldown-period-for-dependency-updates