Introduce Artist.of_the_day
This method returns a consistent featured artist for each day of the
year cycling between the possible features artists. This is preferable
in my mind to randomly selecting a featured artists as it provides a
fairer balance and means that the homepage looks the same on a given
day for anyone who visits it.