Authorize access to jobs dashboard w/ User#admin?
Instead of the built-in HTTP Basic auth mechanism.

I've created an `AdminController` so I can set the
`mission_control.jobs.base_controller_class` config option to use it. It
makes use of a somewhat dummy `AdminPolicy` which seems to be the
easiest way to satisfy the `verify_authorized` after action defined in
`ApplicationController`.

I haven't retro-fitted this as the base class controller for
`Admin::NewslettersController`, because that's already protected by
the `NewsletterPolicy` at the model level.