Fix bugs when access to jobs dashboard not allowed
Previously if the user was not signed in and you tried to access
`/admin/jobs`, we saw `ActionController::UrlGenerationError` exceptions [1]
in `ApplicationController#authenticate` when calling the `#log_in_path`
route helper method.

This was being caused by the `ApplicationController` in
`mission_patch-jobs` adding a `server_id` param in its
`#default_url_options` method [2].

A similar problem existed if the user was signed in but was not an
admin. This time the exception was raised in the `#user_not_authorized`
method in `ApplicationController` when calling the `#root_path` route
helper method.

I've fixed both of these problems by calling the route helper methods on
`main_app` [3] as suggested in this comment [4]. I'm not convinced this
is a great solution, but it will do for now!

[1]: https://app.rollbar.com/a/gofreerange/fix/item/jam/147
[2]: https://github.com/rails/mission_control-jobs/blob/9fa76f054c38b1e24b5b8d2c1c63b513ed87c452/app/controllers/mission_control/jobs/application_controller.rb#L20-L22
[3]: https://api.rubyonrails.org/v8.1.1/classes/Rails/Engine.html#class-Rails::Engine-label-Using+Engine-27s+routes+outside+Engine
[4]: https://github.com/rails/mission_control-jobs/issues/245#issuecomment-2906905748