Fix Rails/StrongParametersExpect auto-correct
The unsafe auto-correct for this cop caused one of the tests for
`AlbumsController#create` to fail, because the `tracks_attributes` were
not in the expected format in the params and so they were not being
included which meant that no tracks were being created.

We should probably report this to rubocop - presumably the auto-correct
doesn't handle nested params correctly?