Avoid N+1 query in AlbumsController#show
When the current user has purchased the album, we display the
corresponding downloads.

This commit enhances the existing test to assert that the downloads are
displayed. I had to fix the test by calling `perform_enqueued_jobs` so
that `Purchase#create_purchase_downloads` is called by the
`after_create` callback to make these assertions succeed.

Making the above change triggered a Prosopite exception, because of an
N+1 query, which is fixed by adding the `with_attached_file` [1] call in
the `app/views/albums/show.html.erb` template.

[1]: https://api.rubyonrails.org/v8.1.1/classes/ActiveStorage/Attached/Model.html#method-i-with_attached_-2A