Allow users to add a catalogue number to a Release
This commit allows users to add an optional catalogue number to a
release. We're only surfacing the number on the Album show page for
now, but this feature was suggested by one of our label beta-testers
who said it helps keep data consistent between various systems they
use.

The 12 character limit on catalogue numbers seems generous enough
given the catalogue numbers used in the wild[1].

[1] e.g. https://en.wikipedia.org/wiki/Catalog_number_(music)