Run development postgres inside docker
I've recently started using linux for development, and it seems like a
relatively straightforward way of running multiple versions of
postgres on the same machine (where I'd previously use homebrew on OS
X) is to run them inside a docker container.

Running the rest of the rails app on the system has some
advantages. I followed a tutorial[1] to get this working, but it's
required changing
the host to an explicit `127.0.0.1` from `localhost`. I'm hoping this shouldn't cause
issues for other developers.

[1] https://www.codewithjason.com/dockerize-rails-apps-database/